Fixed locking problems

+++ gldcore/lock.h
@@ -19,7 +19,7 @@ extern "C" {
 
 #include <sys/types.h>
 
-typedef int64_t LOCKVAR;
+typedef int LOCKVAR;
 void rlock(LOCKVAR *lock);
 void wlock(LOCKVAR *lock);
 void runlock(LOCKVAR *lock);
